Proves varies de funcionament (II)
Corregit un bug que feia que el v_fix quan s'evalués ho fes només 
tenint en compte el primer dels 3 parametres, falta que es printin 
bé encara.

Segmentation faults en 3 funcions, ja que he arreglat el problema amb 
les lògiques.

Problema llegint X i Y

+++ main.cpp
@@ -35,7 +35,7 @@ Node* root =  read ( sin);
     
   for (i = 0; i < m; i++) {
     for (j = 0; j < n; j++) {
-      Env e(i/double(m),j/double(n));
+      Env e(double(i)/double(m),double(j)/double(n));
       imatge[i][j] = root->eval(e);
       
     }
